+++ pyvela/scripts/par_tim-to-jlso.py
@@ -1,9 +1,0 @@
-#!/usr/bin/env python
-
-import sys
-
-from pyvela.spnta import par_tim_to_jlso
-
-parfile, timfile, jlsofile = sys.argv[1], sys.argv[2], sys.argv[3]
-
-par_tim_to_jlso(parfile, timfile, jlsofile)